Rename EmailClient -> SESEmailClient, IEmailClient -> EmailClient
