+++ service/src/main/kotlin/fi/espoo/evaka/reservations/MissingReservationsReminders.kt
@@ -7,7 +7,7 @@ package fi.espoo.evaka.reservations
 import fi.espoo.evaka.EmailEnv
 import fi.espoo.evaka.daycare.domain.Language
 import fi.espoo.evaka.emailclient.Email
-import fi.espoo.evaka.emailclient.IEmailClient
+import fi.espoo.evaka.emailclient.EmailClient
 import fi.espoo.evaka.emailclient.IEmailMessageProvider
 import fi.espoo.evaka.pis.EmailMessageType
 import fi.espoo.evaka.placement.PlacementType
@@ -29,7 +29,7 @@ import org.springframework.stereotype.Service
 class MissingReservationsReminders(
     private val featureConfig: FeatureConfig,
     private val asyncJobRunner: AsyncJobRunner<AsyncJob>,
-    private val emailClient: IEmailClient,
+    private val emailClient: EmailClient,
     private val emailMessageProvider: IEmailMessageProvider,
     private val emailEnv: EmailEnv
 ) {
